edits

+++ models/assignment_work/staging/stg_nyc_open_restaurant_apps.sql
@@ -1,13 +1,11 @@
 -- Clean and standardize NYC Open Restaurant Applications data
--- One row per application
-
 WITH source AS (
     SELECT * FROM {{ source('raw', 'source_nyc_open_restaurant_apps') }}
 ),
 
 cleaned AS (
     SELECT
-        -- Get all columns from source, except ones we're transforming below
+        -- 1. Exclude the raw columns we are about to transform
         * EXCEPT (
             unique_key,
             time_of_submission,
@@ -24,16 +22,13 @@ cleaned AS (
             sidewalk_dimensions_width
         ),
 
-        -- Identifiers
+        -- 2. Identifiers & Timestamps
         CAST(unique_key AS STRING) AS application_id,
-
-        -- Date/Time
         CAST(time_of_submission AS TIMESTAMP) AS submitted_at,
 
-        -- Location - address details
+        -- 3. Location cleaning (Fixing the 'bulding' typo from source)
         CAST(bulding_number AS STRING) AS building_number,
         
-        -- Location - clean zip code
         CASE
             WHEN UPPER(TRIM(CAST(zip AS STRING))) IN ('N/A', 'NA') THEN NULL
             WHEN UPPER(TRIM(CAST(zip AS STRING))) = 'ANONYMOUS' THEN 'Anonymous'
@@ -45,7 +40,6 @@ cleaned AS (
             ELSE NULL
         END AS zip_code,
 
-        -- Location - standardized borough
         CASE
             WHEN UPPER(TRIM(borough)) IN ('MANHATTAN', 'NEW YORK COUNTY') THEN 'Manhattan'
             WHEN UPPER(TRIM(borough)) IN ('BRONX', 'THE BRONX') THEN 'Bronx'
@@ -55,30 +49,28 @@ cleaned AS (
             ELSE 'UNKNOWN or CITYWIDE'
         END AS borough,
 
-        CAST(latitude AS DECIMAL) AS latitude,
-        CAST(longitude AS DECIMAL) AS longitude,
+        CAST(latitude AS FLOAT64) AS latitude,
+        CAST(longitude AS FLOAT64) AS longitude,
 
-        -- Dimensions (Casting to Decimal for calculations)
-        CAST(roadway_dimensions_area AS DECIMAL) AS roadway_area,
-        CAST(roadway_dimensions_length AS DECIMAL) AS roadway_length,
-        CAST(roadway_dimensions_width AS DECIMAL) AS roadway_width,
-        CAST(sidewalk_dimensions_area AS DECIMAL) AS sidewalk_area,
-        CAST(sidewalk_dimensions_length AS DECIMAL) AS sidewalk_length,
-        CAST(sidewalk_dimensions_width AS DECIMAL) AS sidewalk_width,
+        -- 4. Dimensions
+        CAST(roadway_dimensions_area AS FLOAT64) AS roadway_area,
+        CAST(roadway_dimensions_length AS FLOAT64) AS roadway_length,
+        CAST(roadway_dimensions_width AS FLOAT64) AS roadway_width,
+        CAST(sidewalk_dimensions_area AS FLOAT64) AS sidewalk_area,
+        CAST(sidewalk_dimensions_length AS FLOAT64) AS sidewalk_length,
+        CAST(sidewalk_dimensions_width AS FLOAT64) AS sidewalk_width,
 
         -- Metadata
         CURRENT_TIMESTAMP() AS _stg_loaded_at
 
     FROM source
 
-    -- Filters
-    -- Note: WHERE can still see the original names because it runs BEFORE the SELECT
+    -- WHERE runs BEFORE SELECT, so it still uses the original source names
     WHERE unique_key IS NOT NULL
-    AND time_of_submission IS NOT NULL
-    AND CAST(time_of_submission AS DATE) >= DATE_SUB(CURRENT_DATE(), INTERVAL 7 YEAR)
+      AND time_of_submission IS NOT NULL
+      AND CAST(time_of_submission AS DATE) >= DATE_SUB(CURRENT_DATE(), INTERVAL 7 YEAR)
 
-    -- Deduplicate
-    -- Note: QUALIFY must use the NEW ALIASES because it runs AFTER the SELECT
+    -- QUALIFY runs AFTER SELECT, so it MUST use your new aliases
     QUALIFY ROW_NUMBER() OVER (PARTITION BY application_id ORDER BY submitted_at DESC) = 1
 )
